Bit of a tidy up of OMEGA_THROW and fixed a bug in string_t::Format()

+++ src/OOCore/Types.cpp
@@ -235,10 +235,10 @@ OMEGA_DEFINE_EXPORTED_FUNCTION(void*,string_t_mid,3,((in),const void*,s1,(in),si
 
 OMEGA_DEFINE_EXPORTED_FUNCTION(void*,string_t_format,2,((in),const Omega::char_t*,sz,(in),va_list,ap))
 {
-	size_t len2 = 0;
-	for (size_t len = 64;;len *= 2)
+	int len2 = 0;
+	for (int len = 64;;len *= 2)
 	{
-		const ACE_String_Base<Omega::char_t> s(len);
+		const ACE_String_Base<Omega::char_t> s(static_cast<size_t>(len));
 		
 		if ((len2 = ACE_OS::vsnprintf(const_cast<Omega::char_t*>(s.fast_rep()),len,sz,ap)) >=0)
 		{
